RDD, Spark Basics, and SparkSQL
Spark operates in [Resilient Distributed Datasets](http://spark.apache.org/docs/latest/programming-guide.html#resilient-distributed-datasets-rdds) (RDDs). An RDD is a collection of data partitioned across machines. RDDs allow the processing of data to be parallelized due to the partitions. RDDs can be created from a SparkContext in two ways: loading an external dataset, or by parallelizing an existing collection of objects in your currently running program (in our Python programs, this is often times a list). In real-life scenario where some data manipulation/cleaning is required before we can query the data with SparkSQL. We will be using a dataset of user information and a data set of purchases that our users have made. We'll be cleaning the data in a regular Spark RDD before querying it with SparkSQL.